chore: 탐색 상세 > 데이터 리니지 - 캔버스 영역 선택시, 미리보기 닫힘버튼 구현코드 이동
